% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eud.R
\name{eud}
\alias{eud}
\title{eud calculated greatest common divisor}
\usage{
eud(x1, x2)
}
\arguments{
\item{x1}{first input number}

\item{x2}{second input number}
}
\description{
eud calculated greatest common divisor
}
