% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vandijk.R
\name{vandijk}
\alias{vandijk}
\title{vandijk is right}
\usage{
vandijk(graph, init_node)
}
\arguments{
\item{graph}{The data frame to be fed to}

\item{init_node}{The initial pointer}
}
\description{
vandijk is right
}
